PoC: forward guest notifications to client
On Linux, applications can issue notifications by invoking Notify method
on DBus object at /org/freedesktop/Notifications.

Listen for these method invocations, extract some notifications details
and forward them to the client.

For monitoring purposes, a new "private" DBus connection is established,
i.e. one that is not shared with any other parts of the code in vdagent.
This is unnecessary at the moment, however, it ensures that GDBus will
behave as expected if any other part of the vdagent starts using it
in the future.

g_dbus_connection_add_filter() is used as a callback for the eavesdropped
Notify method calls. The DBus messages are already filtered by the MATCH_RULE
that is set in the start_monitoring().
The filter function also consumes all DBus messages by returning NULL.
This ensures that GDBus won't internally respond to any of them.

The filter function (notify_invoked_cb) is called from GDBusWorkerThread.
But since udscs_connection cannot be used from multiple threads,
the GDBusMessage needs to be transfered to the main thread.
This is done by adding the message to a GAsyncQueue. The main thread
is subsequently notified about this using g_idle_add().